requirements.txt
uvicorn
fastapi
prisma
pyjwt
passlib
bcrypt
pydantic[email]
python-dotenv
cryptography
